fix: add job url to comment
